requirements.txt
streamlit==1.41.1
PyPDF2==3.0.1
openai==1.57.4
langchain-openai==0.2.12
nest-asyncio==1.6.0
pinecone-client==5.0.1
langchain==0.3.12
langchain-pinecone==0.2.0